+++ integration_tests/test_load_dxf_unicode_notation.py
@@ -2,12 +2,13 @@
 #  License: MIT License
 import os
 import pytest
-import ezdxf
+from ezdxf import recover
+
 BASEDIR = os.path.dirname(__file__)
 DATADIR = 'data'
 
 
-@pytest.fixture(params=['Leica_Disto_S910.dxf'])
+@pytest.fixture(params=['ASCII_R12.dxf'])
 def filename(request):
     filename = os.path.join(BASEDIR, DATADIR, request.param)
     if not os.path.exists(filename):
@@ -15,9 +16,11 @@ def filename(request):
     return filename
 
 
-def test_leica_disto_r12(filename):
-    doc = ezdxf.readfile(filename)
+def test_load_special_dxf_unicode_notation(filename):
+    doc, auditor = recover.readfile(filename)
+    layer = doc.layers.get('ΛΑΓΕΡÄÜÖ')
+    assert layer.dxf.name == 'ΛΑΓΕΡÄÜÖ'
+
     msp = doc.modelspace()
-    points = list(msp.query('POINT'))
-    assert len(points) == 11
-    assert len(points[0].dxf.location) == 3
+    lines = msp.query('LINE[layer=="ΛΑΓΕΡÄÜÖ"]')
+    assert len(lines) == 2